Finished Upload Data Screen, complete with error messages and full functionality. One Error will need to be modified but need to speak to team about it first. Changed heart rate from a double to an int. Made some changes to Parser, Activity, DateTime, Entry and Position to get it working.

+++ src/main/java/seng202/team10/Control/DataAnalysis.java
@@ -76,9 +76,9 @@ public class DataAnalysis {
         return distanceArray;
     }
 
-    public ArrayList<Double> getHeartRateFromActivity(Activity activity) {
+    public ArrayList<Integer> getHeartRateFromActivity(Activity activity) {
 
-        ArrayList<Double> heartRateArray = new ArrayList<>();
+        ArrayList<Integer> heartRateArray = new ArrayList<>();
         for (Entry entry : activity.getEntries()) {
             heartRateArray.add(entry.getHeartRate());
         }
